backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
pydantic==2.5.3
python-multipart==0.0.6
apscheduler==3.10.4
redis==5.0.1
python-dotenv==1.0.0

# OpenTelemetry for OTLP logging
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-http>=1.21.0

# Datadog APM tracing (agentless)
ddtrace>=2.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
httpx==0.26.0
